Deleted an unaccessible module.
